Privcand: Created db at first RPC instead of socket creation, fix discard/delete Privcand: On non-conflict, merge at update Change int to size_t in cxvec_append and related functions

+++ lib/clixon/clixon_xml_map.h
@@ -63,9 +63,9 @@ int xmlns_assign(cxobj *x);
 int xml2cvec(cxobj *xt, yang_stmt *ys, cvec **cvv0);
 int cvec2xml_1(cvec *cvv, char *toptag, cxobj *xp, cxobj **xt0);
 int xml_diff(cxobj *x0, cxobj *x1,
-             cxobj ***first, int *firstlen,
-             cxobj ***second, int *secondlen,
-             cxobj ***changed_x0, cxobj ***changed_x1, int *changedlen);
+             cxobj ***first, size_t *firstlen,
+             cxobj ***second, size_t *secondlen,
+             cxobj ***changed_x0, cxobj ***changed_x1, size_t *changedlen);
 int xml_tree_equal(cxobj *x0, cxobj *x1);
 int xml_tree_prune_flagged_sub(cxobj *xt, int flag, int test, int *upmark);
 int xml_tree_mark_flagged_sub(cxobj *xt, int flag, int test, int delmark, int *upmark);
